# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/jobs:
    get:
      operationId: adminListJobs
      tags: [careers]
      summary: List all job postings including closed/archived (admin only)
      responses:
        "200":
          description: All jobs, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
    post:
      operationId: adminCreateJob
      tags: [careers]
      summary: Create a job posting (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid job
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
  /admin/jobs/{id}:
    patch:
      operationId: adminUpdateJob
      tags: [careers]
      summary: Update a job posting (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid update
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Job not found
    delete:
      operationId: adminDeleteJob
      tags: [careers]
      summary: Delete a job posting (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Job not found
  /careers/resume-upload-url:
    post:
      operationId: requestResumeUploadUrl
      tags: [careers]
      summary: Request a presigned upload URL for a resume (public, rate limited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUploadRequest"
      responses:
        "200":
          description: Presigned upload URL and internal object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeUploadResponse"
        "400":
          description: Unsupported file type or size
        "429":
          description: Rate limited
  /careers/applications:
    post:
      operationId: submitJobApplication
      tags: [careers]
      summary: Submit a job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobApplicationInput"
      responses:
        "201":
          description: Application accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryStatus"
        "400":
          description: Invalid application
        "429":
          description: Rate limited
  /admin/applications:
    get:
      operationId: adminListApplications
      tags: [careers]
      summary: List all job applications (admin only)
      responses:
        "200":
          description: Applications, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobApplication"
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
  /admin/applications/{id}:
    patch:
      operationId: adminUpdateApplication
      tags: [careers]
      summary: Update an application's status, recruiter, or notes (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationRequest"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
        "400":
          description: Invalid update
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Application not found
    delete:
      operationId: adminDeleteApplication
      tags: [careers]
      summary: Delete an application (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "403":
          description: Authenticated but not an admin
        "404":
          description: Application not found
...
components:
  schemas:
...
    InquiryStatus:
      type: object
      properties:
        status:
          type: string
          enum: [received]
        referenceNumber:
          type: string
          description: Public lead reference number shown to the visitor and used in emails.
      required:
        - status
    Job:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        department: { type: string }
        experience: { type: string }
        employmentType: { type: string }
        location: { type: string }
        workMode: { type: string }
        salary: { type: ["string", "null"] }
        description: { type: string }
        requirements: { type: ["string", "null"] }
        responsibilities: { type: ["string", "null"] }
        skills: { type: ["string", "null"] }
        status: { type: string, enum: [open, closed, archived] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        [
          id,
          title,
          department,
          experience,
          employmentType,
          location,
          workMode,
          salary,
          description,
          requirements,
          responsibilities,
          skills,
          status,
          createdAt,
          updatedAt,
        ]
    JobInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        department: { type: string, minLength: 1 }
        experience: { type: string, minLength: 1 }
        employmentType: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        workMode: { type: string, minLength: 1 }
        salary: { type: ["string", "null"] }
        description: { type: string, minLength: 1 }
        requirements: { type: ["string", "null"] }
        responsibilities: { type: ["string", "null"] }
        skills: { type: ["string", "null"] }
        status: { type: string, enum: [open, closed, archived] }
      required:
        [title, department, experience, employmentType, location, workMode, description]
    JobUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        department: { type: string, minLength: 1 }
        experience: { type: string, minLength: 1 }
        employmentType: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        workMode: { type: string, minLength: 1 }
        salary: { type: ["string", "null"] }
        description: { type: string, minLength: 1 }
        requirements: { type: ["string", "null"] }
        responsibilities: { type: ["string", "null"] }
        skills: { type: ["string", "null"] }
        status: { type: string, enum: [open, closed, archived] }
    JobApplicationInput:
      type: object
      properties:
        jobId: { type: ["integer", "null"] }
        fullName: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        phone: { type: string, minLength: 5 }
        city: { type: string, minLength: 1 }
        qualification: { type: string, minLength: 1 }
        college: { type: string }
        graduationYear: { type: string }
        experience: { type: string }
        skills: { type: string, minLength: 1 }
        linkedin: { type: string }
        github: { type: string }
        portfolio: { type: string }
        preferredRole: { type: string, minLength: 1 }
        expectedSalary: { type: string }
        joiningAvailability: { type: string }
        resumePath: { type: string, minLength: 1 }
        coverLetter: { type: string }
        website:
          type: string
          description: Honeypot field — must be empty; bots that fill it are silently dropped.
      required:
        [fullName, email, phone, city, qualification, skills, preferredRole, resumePath]
    JobApplication:
      type: object
      properties:
        id: { type: integer }
        referenceNumber: { type: string }
        jobId: { type: ["integer", "null"] }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        city: { type: string }
        qualification: { type: string }
        college: { type: ["string", "null"] }
        graduationYear: { type: ["string", "null"] }
        experience: { type: ["string", "null"] }
        skills: { type: string }
        linkedin: { type: ["string", "null"] }
        github: { type: ["string", "null"] }
        portfolio: { type: ["string", "null"] }
        preferredRole: { type: string }
        expectedSalary: { type: ["string", "null"] }
        joiningAvailability: { type: ["string", "null"] }
        resumePath: { type: string }
        coverLetter: { type: ["string", "null"] }
        status:
          type: string
          enum: [new, shortlisted, interview_scheduled, selected, rejected]
        assignedRecruiter: { type: ["string", "null"] }
        internalNotes: { type: ["string", "null"] }
        ipAddress: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        [
          id,
          referenceNumber,
          jobId,
          fullName,
          email,
          phone,
          city,
          qualification,
          college,
          graduationYear,
          experience,
          skills,
          linkedin,
          github,
          portfolio,
          preferredRole,
          expectedSalary,
          joiningAvailability,
          resumePath,
          coverLetter,
          status,
          assignedRecruiter,
          internalNotes,
          ipAddress,
          createdAt,
          updatedAt,
        ]
    ResumeUploadRequest:
      type: object
      properties:
        fileName: { type: string, minLength: 1 }
        fileSize: { type: integer, minimum: 1 }
      required: [fileName, fileSize]
    ResumeUploadResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]
    UpdateApplicationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [new, shortlisted, interview_scheduled, selected, rejected]
        assignedRecruiter: { type: ["string", "null"] }
        internalNotes: { type: ["string", "null"] }
...